requirements.txt
# Interactive Brokers API
ib_insync

# Data manipulation
pandas
numpy

# Utilities
python-dotenv
pyyaml

# Logging
loguru

# Testing
pytest
pytest-asyncio

# Web Dashboard (Streamlit is installed, fastapi failed)
# fastapi
# uvicorn[standard]
# websockets
# pydantic

# Data storage
# pyarrow

# Optional: Technical indicators
ta

# Machine Learning
scikit-learn
xgboost
lightgbm
optuna
imbalanced-learn
joblib
matplotlib
seaborn
plotly
